LPS-78208 Move xyz.handler. to "xyz" for *CriterionHandler
