+++ modules/apps/collaboration/item-selector/item-selector-criteria-impl/src/main/java/com/liferay/item/selector/criteria/internal/file/FileItemSelectorCriterionHandler.java
@@ -12,32 +12,39 @@
  * details.
  */
 
-package com.liferay.journal.item.selector.web.internal.handler;
+package com.liferay.item.selector.criteria.internal.file;
 
 import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
 import com.liferay.item.selector.ItemSelectorCriterionHandler;
-import com.liferay.journal.item.selector.criterion.JournalItemSelectorCriterion;
+import com.liferay.item.selector.criteria.file.criterion.FileItemSelectorCriterion;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 
 /**
- * @author Eduardo Garcia
+ * @author Roberto Díaz
  */
 @Component(service = ItemSelectorCriterionHandler.class)
-public class JournalItemSelectorCriterionHandler
-	extends BaseItemSelectorCriterionHandler<JournalItemSelectorCriterion> {
+public class FileItemSelectorCriterionHandler
+	extends BaseItemSelectorCriterionHandler<FileItemSelectorCriterion> {
 
 	@Override
-	public Class<JournalItemSelectorCriterion> getItemSelectorCriterionClass() {
-		return JournalItemSelectorCriterion.class;
+	public Class<FileItemSelectorCriterion> getItemSelectorCriterionClass() {
+		return FileItemSelectorCriterion.class;
 	}
 
 	@Activate
 	@Override
 	protected void activate(BundleContext bundleContext) {
 		super.activate(bundleContext);
+	}
+
+	@Deactivate
+	@Override
+	protected void deactivate() {
+		super.deactivate();
 	}
 
 }
