+++ modules/apps/collaboration/item-selector/item-selector-criteria-impl/src/main/java/com/liferay/item/selector/criteria/internal/image/ImageItemSelectorCriterionHandler.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.user.groups.admin.item.selector.web.internal.handler;
+package com.liferay.item.selector.criteria.internal.image;
 
 import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
 import com.liferay.item.selector.ItemSelectorCriterionHandler;
-import com.liferay.user.groups.admin.item.selector.UserGroupItemSelectorCriterion;
+import com.liferay.item.selector.criteria.image.criterion.ImageItemSelectorCriterion;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
@@ -24,17 +24,15 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
 
 /**
- * @author Alessio Antonio Rendina
+ * @author Roberto Díaz
  */
 @Component(service = ItemSelectorCriterionHandler.class)
-public class UserGroupItemSelectorCriterionHandler
-	extends BaseItemSelectorCriterionHandler<UserGroupItemSelectorCriterion> {
+public class ImageItemSelectorCriterionHandler
+	extends BaseItemSelectorCriterionHandler<ImageItemSelectorCriterion> {
 
 	@Override
-	public Class<UserGroupItemSelectorCriterion>
-		getItemSelectorCriterionClass() {
-
-		return UserGroupItemSelectorCriterion.class;
+	public Class<ImageItemSelectorCriterion> getItemSelectorCriterionClass() {
+		return ImageItemSelectorCriterion.class;
 	}
 
 	@Activate
