+++ modules/apps/collaboration/item-selector/item-selector-criteria-impl/src/main/java/com/liferay/item/selector/criteria/internal/upload/UploadItemSelectorCriterionHandler.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.knowledge.base.item.selector.web.internal.handler;
+package com.liferay.item.selector.criteria.internal.upload;
 
 import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
 import com.liferay.item.selector.ItemSelectorCriterionHandler;
-import com.liferay.knowledge.base.item.selector.criterion.KBAttachmentItemSelectorCriterion;
+import com.liferay.item.selector.criteria.upload.criterion.UploadItemSelectorCriterion;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
@@ -24,18 +24,15 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
 
 /**
- * @author Roberto Díaz
+ * @author Ambrín Chaudhary
  */
 @Component(service = ItemSelectorCriterionHandler.class)
-public class KBAttachmentItemSelectorCriterionHandler
-	extends BaseItemSelectorCriterionHandler
-		<KBAttachmentItemSelectorCriterion> {
+public class UploadItemSelectorCriterionHandler
+	extends BaseItemSelectorCriterionHandler<UploadItemSelectorCriterion> {
 
 	@Override
-	public Class<KBAttachmentItemSelectorCriterion>
-		getItemSelectorCriterionClass() {
-
-		return KBAttachmentItemSelectorCriterion.class;
+	public Class<UploadItemSelectorCriterion> getItemSelectorCriterionClass() {
+		return UploadItemSelectorCriterion.class;
 	}
 
 	@Activate
