+++ modules/apps/collaboration/wiki/wiki-service/src/main/java/com/liferay/wiki/internal/item/selector/WikiAttachmentItemSelectorCriterionHandler.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.blogs.item.selector.web.internal.handler;
+package com.liferay.wiki.internal.item.selector;
 
-import com.liferay.blogs.item.selector.criterion.BlogsItemSelectorCriterion;
 import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
 import com.liferay.item.selector.ItemSelectorCriterionHandler;
+import com.liferay.wiki.item.selector.criterion.WikiAttachmentItemSelectorCriterion;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
@@ -24,15 +24,18 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
 
 /**
- * @author Roberto Díaz
+ * @author Iván Zaera
  */
 @Component(service = ItemSelectorCriterionHandler.class)
-public class BlogsItemSelectorCriterionHandler
-	extends BaseItemSelectorCriterionHandler<BlogsItemSelectorCriterion> {
+public class WikiAttachmentItemSelectorCriterionHandler
+	extends BaseItemSelectorCriterionHandler
+		<WikiAttachmentItemSelectorCriterion> {
 
 	@Override
-	public Class<BlogsItemSelectorCriterion> getItemSelectorCriterionClass() {
-		return BlogsItemSelectorCriterion.class;
+	public Class<WikiAttachmentItemSelectorCriterion>
+		getItemSelectorCriterionClass() {
+
+		return WikiAttachmentItemSelectorCriterion.class;
 	}
 
 	@Activate
