+++ modules/apps/web-experience/journal/journal-item-selector-web/src/main/java/com/liferay/journal/item/selector/web/internal/JournalItemSelectorCriterionHandler.java
@@ -12,39 +12,32 @@
  * details.
  */
 
-package com.liferay.item.selector.criteria.internal.url.handler;
+package com.liferay.journal.item.selector.web.internal;
 
 import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
 import com.liferay.item.selector.ItemSelectorCriterionHandler;
-import com.liferay.item.selector.criteria.url.criterion.URLItemSelectorCriterion;
+import com.liferay.journal.item.selector.criterion.JournalItemSelectorCriterion;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 
 /**
- * @author Roberto Díaz
+ * @author Eduardo Garcia
  */
 @Component(service = ItemSelectorCriterionHandler.class)
-public class URLItemSelectorCriterionHandler
-	extends BaseItemSelectorCriterionHandler<URLItemSelectorCriterion> {
+public class JournalItemSelectorCriterionHandler
+	extends BaseItemSelectorCriterionHandler<JournalItemSelectorCriterion> {
 
 	@Override
-	public Class<URLItemSelectorCriterion> getItemSelectorCriterionClass() {
-		return URLItemSelectorCriterion.class;
+	public Class<JournalItemSelectorCriterion> getItemSelectorCriterionClass() {
+		return JournalItemSelectorCriterion.class;
 	}
 
 	@Activate
 	@Override
 	protected void activate(BundleContext bundleContext) {
 		super.activate(bundleContext);
-	}
-
-	@Deactivate
-	@Override
-	protected void deactivate() {
-		super.deactivate();
 	}
 
 }
